+++ src/lib/components/features/category/index.ts
@@ -1,4 +1,5 @@
 export { default as CategoryArchive } from './category-archive.svelte';
+export { default as CategoryArchivedNotice } from './category-archived-notice.svelte';
 export { default as CategoryCreate } from './category-create.svelte';
 export { default as CategoryDelete } from './category-delete.svelte';
 export { default as CategoryEdit } from './category-edit.svelte';